requirements.txt
Flask==2.1.0
gspread==5.7.0
oauth2client==4.1.3
pandas==1.5.3
line-bot-sdk==2.4.2
gunicorn==20.1.0
python-dotenv==0.21.1
requests==2.28.2
Werkzeug==2.2.3
numpy==1.24.3